+++ src/main/java/com/mageddo/dnsproxyserver/config/application/ConfigPathMapper.java
@@ -1,15 +1,17 @@
-package com.mageddo.dnsproxyserver.config.dataprovider;
+package com.mageddo.dnsproxyserver.config.application;
 
+import com.mageddo.dnsproxyserver.config.dataprovider.vo.ConfigFlag;
 import com.mageddo.utils.Files;
 import com.mageddo.utils.Runtime;
+import com.mageddo.utils.Tests;
 import lombok.extern.slf4j.Slf4j;
 
 import java.nio.file.Path;
 
 @Slf4j
-public class ConfigPathBuilder {
+class ConfigPathMapper {
   private static Path build0(Path workDir, Path configPath) {
-    if (ConfigDAOJson.runningInTestsAndNoCustomConfigPath(configPath)) {
+    if (runningInTestsAndNoCustomConfigPath(configPath)) {
       final var file = Files.createTempFileDeleteOnExit("dns-proxy-server-junit", ".json");
       log.trace("status=runningInTests, usingEmptyFile={}", file);
       return file;
@@ -33,5 +35,13 @@ public class ConfigPathBuilder {
     final var path = build0(workDir, configFilePath);
     log.debug("status=configPathBuilt, path={}", path);
     return path;
+  }
+
+  private static boolean runningInTestsAndNoCustomConfigPath(Path configPath) {
+    return isDefaultConfigFilePath(configPath) && Tests.inTest();
+  }
+
+  private static boolean isDefaultConfigFilePath(Path configPath) {
+    return ConfigFlag.DEFAULT_CONFIG_FILE_AS_PATH.equals(configPath);
   }
 }
