+++ src/main/java/com/mageddo/dnsproxyserver/config/dataprovider/ConfigDAOJson.java
@@ -3,8 +3,6 @@ package com.mageddo.dnsproxyserver.config.dataprovider;
 import com.mageddo.dnsproxyserver.config.Config;
 import com.mageddo.dnsproxyserver.config.application.ConfigFileFinderService;
 import com.mageddo.dnsproxyserver.config.dataprovider.mapper.ConfigJsonV2Mapper;
-import com.mageddo.dnsproxyserver.config.dataprovider.vo.ConfigFlag;
-import com.mageddo.utils.Tests;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -28,14 +26,6 @@ public class ConfigDAOJson implements ConfigDAO {
     final var jsonConfig = JsonConfigs.loadConfig(configPath);
     log.debug("configPath={}", configPath);
     return ConfigJsonV2Mapper.toConfig(jsonConfig, configPath);
-  }
-
-  static boolean runningInTestsAndNoCustomConfigPath(Path configPath) {
-    return isDefaultConfigFilePath(configPath) && Tests.inTest();
-  }
-
-  private static boolean isDefaultConfigFilePath(Path configPath) {
-    return ConfigFlag.DEFAULT_CONFIG_FILE_AS_PATH.equals(configPath);
   }
 
   @Override
