+++ src/test/java/com/mageddo/dnsproxyserver/config/dataprovider/ConfigDAOJsonTest.java
@@ -6,11 +6,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
 import org.mockito.junit.jupiter.MockitoExtension;
-import testing.templates.ConfigFlagTemplates;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 
 import static com.mageddo.utils.TestUtils.readAndSortJsonExcluding;
 import static com.mageddo.utils.TestUtils.readAsStream;
@@ -27,21 +25,6 @@ class ConfigDAOJsonTest {
   };
 
   final ConfigDAOJson configDAOJson = new ConfigDAOJson(null);
-
-  @Test
-  void mustBuildConfPathRelativeToWorkDir(@TempDir Path tmpDir){
-    // arrange
-    final var workDir = tmpDir.resolve("custom-work-dir");
-    final var flags = ConfigFlagTemplates.defaultWithConfigPath(Paths.get("conf/config-custom.json"));
-    ConfigDAOCmdArgs.setArgs(flags.getArgs());
-
-    // act
-    final var configPath = ConfigPathBuilder.build(workDir, flags.getConfigFileAsPath());
-
-    // assert
-    assertEquals("config-custom.json", configPath.getFileName().toString());
-    assertEquals(workDir.getFileName().toString(), configPath.getParent().getParent().getFileName().toString());
-  }
 
   @Test
   void mustReadAndRespectStoredConfigFile(@TempDir Path tmpDir) {
